Add a way for notification that a PartialFuturable action is completed.
When you wish to get notification of completion you should implement
PartialFuturableCompleted. A simple implementation
"SimplePartialFuturableCompleted" can be used to extend from to provide
this functionality.

Add a way for things to notify invocation handler that all operation
have been completed. Added optimization so that when all operations
from the operations are completed no pending future states are checked.
This allows a partial operation to not have to actually call set on
every field that a consumer may use.